detach listener from pubsub streamer
